refactor(pm): expose command handlers directly
