+++ crates/pm/src/main.rs
@@ -1,3 +1,3 @@
 fn main() {
-    utoo_pm::cli_main();
+    utoo_pm::run_cli();
 }
